src/backend/requirements.txt
fastapi
uvicorn
python-dotenv
robotframework
robotframework-seleniumlibrary
docker
pydantic
pydantic-settings
PyYAML
ollama
crewai
crewai-tools
langchain-google-genai
beautifulsoup4
selenium
webdriver-manager
sse-starlette
aiohttp
